+++ hoomd/md/TableAngleForceGPU.cu
@@ -38,7 +38,7 @@
 */
 __global__ void gpu_compute_table_angle_forces_kernel(Scalar4* d_force,
                                      Scalar* d_virial,
-                                     const unsigned int virial_pitch,
+                                     const size_t virial_pitch,
                                      const unsigned int N,
                                      const Scalar4 *d_pos,
                                      const BoxDim box,
@@ -239,7 +239,7 @@ __global__ void gpu_compute_table_angle_forces_kernel(Scalar4* d_force,
 */
 hipError_t gpu_compute_table_angle_forces(Scalar4* d_force,
                                      Scalar* d_virial,
-                                     const unsigned int virial_pitch,
+                                     const size_t virial_pitch,
                                      const unsigned int N,
                                      const Scalar4 *d_pos,
                                      const BoxDim &box,
